requirements.txt
Jinja2==2.7.3
Markdown==2.5.2
MarkupSafe==0.23
PyYAML==3.11
Pygments==2.0.2
Unidecode==0.04.17
python-slugify==0.1.0
wsgiref==0.1.2
